Refactored and documented AbstractEntity. Begun implementation of projectile. Tower now simulates firing.

+++ potatoes/src/main/java/com/deco2800/potatoes/entities/Peon.java
@@ -21,7 +21,7 @@ public class Peon extends AbstractEntity implements Tickable {
 	}
 
 	@Override
-	public void onTick(int i) {
+	public void onTick(long i) {
 
 	}
 }
